+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- Support drawing columnar data `scatterplot.draw({ x: [...], y: [...], ... })` (#59)
+
 ## v0.19.2
 
 - Fix an issue with not properly publishing the pointout event (#58)